Code File
The “4-Queens Problem” consists of placing four queens on a 4 x 4 chessboard so that no two queens can capture each other. That is, no two queens are allowed to be placed on the same row, the same column or the same diagonal (both primary and secondary diagonals).  Write a C program to simulate the given problem and perform the algorithmic complexity analysis for the solution you propose.